requirements.txt
fastapi
uvicorn
websockets
aiohttp
pydantic
python-dotenv